requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
attrs==22.2.0
charset-normalizer==3.1.0
discord.py==2.2.2
frozenlist==1.3.3
idna==3.4
iso8601==1.1.0
multidict==6.0.4
numpy==1.24.2
pandas==1.5.3
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.2
six==1.16.0
twitchio==2.6.0
typing_extensions==4.5.0
yarl==1.8.2
